+++ devenv.nix
@@ -249,6 +249,11 @@ let
       name = "notion-react";
       port = 6014;
     }
+    {
+      path = "packages/@overeng/notion-md";
+      name = "notion-md";
+      port = 6015;
+    }
   ];
   packagesWithNetlifyPreview = lib.filter (pkg: pkg.name != "tui-stories") packagesWithStorybook;
 in